Teste de escuta do socket, onde ele vai retornar se conseguiu fazer a conexao com o "cliente".
